Add main game logic as a callback function
The main game logic is added as a callback function of the event handler
on the grid cells. Helper functions of ui manger handle showing the
marker on the grid.

Reset function is added for resetting the board ui, by removing all the
markers from the board.

start button now starts the game. Soon will be replaced with restart
button and auto starting the game on load.